Stabilize onboarding e2e helpers
