+++ web/tests/e2e/admin-enrollment-approval.spec.ts
@@ -6,186 +6,8 @@ import {
   hasAdminServiceEnv,
   loginAsAdmin,
 } from './helpers/admin-account'
-
-const uniqueSuffix = () => {
-  const now = new Date()
-  return `${now.getFullYear()}${String(now.getMonth() + 1).padStart(2, '0')}${String(now.getDate()).padStart(2, '0')}${String(now.getHours()).padStart(2, '0')}${String(now.getMinutes()).padStart(2, '0')}${String(now.getSeconds()).padStart(2, '0')}${String(now.getMilliseconds()).padStart(3, '0')}`
-}
-
-const fillRequiredFields = async (
-  page: import('@playwright/test').Page,
-  context?: { firstName?: string; surname?: string }
-) => {
-  const form = page.locator('form').first()
-
-  const requiredTextInputs = form.locator(
-    'input[required]:not([type="hidden"]):not([type="radio"]):not([type="checkbox"])'
-  )
-
-  const inputCount = await requiredTextInputs.count()
-  for (let index = 0; index < inputCount; index += 1) {
-    const input = requiredTextInputs.nth(index)
-    const currentValue = await input.inputValue()
-    if (currentValue.trim()) continue
-
-    const inputType = (await input.getAttribute('type')) ?? 'text'
-    const inputName = ((await input.getAttribute('name')) ?? '').toLowerCase()
-
-    if (inputType === 'date') {
-      await input.fill('2012-05-10')
-      continue
-    }
-
-    if (inputType === 'number') {
-      await input.fill('2')
-      continue
-    }
-
-    if (inputType === 'email') {
-      await input.fill(`sai+tests${uniqueSuffix()}@chsolutions.ca`)
-      continue
-    }
-
-    if (inputName.includes('postcode') || inputName.includes('postal')) {
-      await input.fill('K1A 0B1')
-      continue
-    }
-
-    if (inputName.includes('firstname') || inputName.includes('first_name')) {
-      await input.fill(context?.firstName ?? 'Sai')
-      continue
-    }
-
-    if (inputName.includes('surname') || inputName.includes('last_name') || inputName.includes('lastname')) {
-      await input.fill(context?.surname ?? 'tests')
-      continue
-    }
-
-    await input.fill('Auto value')
-  }
-
-  const requiredSelectNames = await form
-    .locator('select[required]')
-    .evaluateAll(selects =>
-      Array.from(new Set(selects.map(select => select.getAttribute('name')).filter((name): name is string => Boolean(name))))
-    )
-
-  for (const name of requiredSelectNames) {
-    const select = form.locator(`select[name="${name}"]`).first()
-    if (!(await select.isVisible())) continue
-    if (!(await select.isEnabled())) continue
-
-    const value = await select.inputValue()
-    if (value) continue
-
-    const options = await select.locator('option').all()
-    for (const option of options) {
-      const optionValue = (await option.getAttribute('value')) ?? ''
-      const disabled = (await option.getAttribute('disabled')) !== null
-      if (!optionValue || disabled) continue
-      await select.selectOption(optionValue)
-      break
-    }
-  }
-
-  const checkboxNames = await form
-    .locator('input[type="checkbox"][name^="question_"], input[type="checkbox"][required]')
-    .evaluateAll(inputs =>
-      Array.from(new Set(inputs.map(input => input.getAttribute('name')).filter((name): name is string => Boolean(name))))
-    )
-
-  for (const name of checkboxNames) {
-    const checkbox = form.locator(`input[type="checkbox"][name="${name}"]`).first()
-    if (!(await checkbox.isVisible())) continue
-    if (!(await checkbox.isEnabled())) continue
-    await checkbox.check()
-  }
-
-  const radioNames = await form
-    .locator('input[type="radio"][required]')
-    .evaluateAll(inputs =>
-      Array.from(new Set(inputs.map(input => input.getAttribute('name')).filter((name): name is string => Boolean(name))))
-    )
-
-  for (const name of radioNames) {
-    const checked = await form.locator(`input[type="radio"][name="${name}"]:checked`).count()
-    if (checked > 0) continue
-    await form.locator(`input[type="radio"][name="${name}"]`).first().check()
-  }
-}
-
-const countMissingRequiredFields = async (page: import('@playwright/test').Page) => {
-  const form = page.locator('form').first()
-  return form.evaluate(formEl => {
-    const getVisible = (element: HTMLElement) => {
-      const style = window.getComputedStyle(element)
-      return style.display !== 'none' && style.visibility !== 'hidden' && element.offsetParent !== null
-    }
-
-    let missing = 0
-
-    const requiredInputs = Array.from(
-      formEl.querySelectorAll<HTMLInputElement>('input[required]:not([type="hidden"])')
-    ).filter(input => !input.disabled && getVisible(input))
-
-    const handledRadioNames = new Set<string>()
-
-    for (const input of requiredInputs) {
-      if (input.type === 'radio') {
-        if (!input.name || handledRadioNames.has(input.name)) continue
-        handledRadioNames.add(input.name)
-        const checked = formEl.querySelector(`input[type="radio"][name="${CSS.escape(input.name)}"]:checked`)
-        if (!checked) missing += 1
-        continue
-      }
-
-      if (input.type === 'checkbox') {
-        if (!input.checked) missing += 1
-        continue
-      }
-
-      if (!input.value.trim()) missing += 1
-    }
-
-    const requiredSelects = Array.from(formEl.querySelectorAll<HTMLSelectElement>('select[required]')).filter(
-      select => !select.disabled && getVisible(select)
-    )
-
-    for (const select of requiredSelects) {
-      if (!select.value) missing += 1
-    }
-
-    return missing
-  })
-}
-
-const submitStepWithRetry = async (
-  page: import('@playwright/test').Page,
-  submitName: RegExp | string,
-  maxAttempts = 3,
-  context?: { firstName?: string; surname?: string }
-) => {
-  for (let attempt = 0; attempt < maxAttempts; attempt += 1) {
-    await fillRequiredFields(page, context)
-
-    const missingBeforeSubmit = await countMissingRequiredFields(page)
-    if (missingBeforeSubmit > 0) {
-      await page.waitForTimeout(400)
-      continue
-    }
-
-    const submitButton = page.getByRole('button', { name: submitName }).first()
-    await submitButton.click()
-    await page.waitForTimeout(500)
-
-    const errorPrompt = page.locator('p.text-red-500').first()
-    if ((await errorPrompt.count()) === 0 || !(await errorPrompt.isVisible())) {
-      return
-    }
-  }
-
-  throw new Error('Form could not be submitted cleanly after retries')
-}
+import { uniqueSuffix } from './helpers/ids'
+import { submitStepWithRetry } from './helpers/onboarding'
 
 const guardianSignupAndRequestEnrollment = async (page: import('@playwright/test').Page) => {
   const suffix = uniqueSuffix()
@@ -223,9 +45,10 @@ const guardianSignupAndRequestEnrollment = async (page: import('@playwright/test
       await page.locator('input[name="additional_guardian_choice"][value="no"]').check()
     }
 
-    await submitStepWithRetry(page, /Save and continue|Saving\.\.\./, 3, {
+    await submitStepWithRetry(page, /Save and continue/i, 8, {
       firstName,
       surname,
+      emailFactory: () => `sai+tests${uniqueSuffix()}@chsolutions.ca`,
     })
     await page.waitForLoadState('networkidle')
   }
